Update Keyence and GPIO to new repo structure
