--- boards/navigation/.cargo/config.toml
@@ -1,9 +1,0 @@
-[target.'cfg(all(target_arch = "arm", target_os = "none"))']
-# replace STM32F429ZITx with your chip as listed in `probe-rs chip list`
-runner = "probe-rs run --chip STM32L476RGTx"
-
-[build]
-target = "thumbv7em-none-eabihf"
-
-[env]
-DEFMT_LOG = "debug,smoltcp=info"
