--- boards/navigation/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-edition = "2021"
-name = "stm-navigation-board"
-version = "0.1.0"
-
-[dependencies]
-embassy-stm32 = { version = "0.1.0", features = ["defmt", "stm32l476rg", "memory-x", "unstable-pac", "time-driver-any", "exti"]  }
-embassy-sync = { version = "0.6.0", features = ["defmt"] }
-embassy-executor = { version = "0.6.0", features = ["task-arena-size-32768", "arch-cortex-m", "executor-thread", "defmt", "integrated-timers"] }
-embassy-time = { version = "0.3.1", features = ["defmt", "defmt-timestamp-uptime", "tick-hz-32_768"] }
-embassy-net = { version = "0.4.0", features = ["defmt", "tcp", "dhcpv4", "medium-ethernet"] }
-embassy-futures = { version = "0.1.0" }
-
-defmt = "0.3"
-defmt-rtt = "0.4"
-
-cortex-m = { version = "0.7.6", features = ["inline-asm", "critical-section-single-core"] }
-cortex-m-rt = "0.7.0"
-embedded-hal = "0.2.6"
-panic-probe = { version = "0.3", features = ["print-defmt"] }
-rand_core = "0.6.3"
-critical-section = "1.1"
-embedded-storage = "0.3.1"
-static_cell = "2"
-
-hyped_sensors = { path = "../../lib/sensors" }
-hyped_io = { path = "../../lib/io" }
-hyped_io_stm32l476rg = { path = "../../lib/io/stm32l476rg" }
-
-[profile.release]
-debug = 2
