--- boards/navigation/build.rs
@@ -1,5 +1,0 @@
-fn main() {
-    println!("cargo:rustc-link-arg-bins=--nmagic");
-    println!("cargo:rustc-link-arg-bins=-Tlink.x");
-    println!("cargo:rustc-link-arg-bins=-Tdefmt.x");
-}
