+++ boards/stm32l476rg/src/io.rs
@@ -1,1 +1,2 @@
+pub mod gpio;
 pub mod i2c;
