+++ boards/stm32l476rg/src/io/gpio.rs
@@ -1,0 +1,19 @@
+use embassy_stm32::gpio::Input;
+use hyped_io::gpio::GpioPin;
+
+/// A GPIO pin on the STM32L476RG.
+pub struct Stm32l476rgGpio {
+    pin: Input<'static>,
+}
+
+impl GpioPin for Stm32l476rgGpio {
+    fn is_high(&mut self) -> bool {
+        self.pin.is_high()
+    }
+}
+
+impl Stm32l476rgGpio {
+    pub fn new(pin: Input<'static>) -> Self {
+        Self { pin }
+    }
+}
