+++ boards/stm32l476rg/src/tasks.rs
@@ -1,1 +1,2 @@
+pub mod keyence;
 pub mod temperature;
