--- lib/io/stm32l476rg/src/lib.rs
@@ -1,3 +1,0 @@
-#![no_std]
-
-pub mod gpio;
